update interface, rename folder, separate assignment and measurement
