+++ db/device_assignments.json
@@ -1,1 +1,1 @@
-{"0": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644724657.1106646"}, "1": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644724794.659023"}, "2": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644724899.0263107"}, "3": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644724935.1753047"}, "4": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644727053.2247057"}, "5": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644727078.4683294"}, "6": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644727128.6137605"}, "7": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727149.9918373"}, "8": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727204.5610213"}, "9": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727246.968356"}, "10": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727292.5712185"}, "11": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727308.5123563"}, "12": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727690.8305037"}, "13": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727950.2577045"}, "14": {"user_id": "5", "device_id": "3", "assigned_by": "4", "assigned_at": "1644731215.2029119"}, "15": {"user_id": "4", "device_id": "3", "assigned_by": "5", "assigned_at": "1644731240.5280757"}, "16": {"user_id": "4", "device_id": "3", "assigned_by": "3", "assigned_at": "1644731258.547831"}, "17": {"user_id": "4", "device_id": "5", "assigned_by": "3", "assigned_at": "1644731285.242353"}, "18": {"user_id": "4", "device_id": "3", "assigned_by": "3", "assigned_at": "1644731296.9164596"}, "19": {"user_id": "5", "device_id": "3", "assigned_by": "4", "assigned_at": "1644731387.2701037"}, "20": {"user_id": "4", "device_id": "3", "assigned_by": "3", "assigned_at": "1644731468.283226"}, "21": {"user_id": "6", "device_id": "3", "assigned_by": "5", "assigned_at": "1644731609.0074632"}, "22": {"user_id": "5", "device_id": "3", "assigned_by": "4", "assigned_at": "1644731854.578424"}, "23": {"user_id": "3", "device_id": "1", "assigned_by": "2", "assigned_at": "1644731877.3139257"}, "24": {"user_id": "5", "device_id": "3", "assigned_by": "4", "assigned_at": "1644731899.8514574"}}
+{"0": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644724657.1106646"}, "1": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644724794.659023"}, "2": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644724899.0263107"}, "3": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644724935.1753047"}, "4": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644727053.2247057"}, "5": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644727078.4683294"}, "6": {"user_id": "1", "device_id": "0", "assigned_by": "0", "assigned_at": "1644727128.6137605"}, "7": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727149.9918373"}, "8": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727204.5610213"}, "9": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727246.968356"}, "10": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727292.5712185"}, "11": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727308.5123563"}, "12": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727690.8305037"}, "13": {"user_id": "1", "device_id": "1", "assigned_by": "0", "assigned_at": "1644727950.2577045"}, "14": {"user_id": "5", "device_id": "3", "assigned_by": "4", "assigned_at": "1644731215.2029119"}, "15": {"user_id": "4", "device_id": "3", "assigned_by": "5", "assigned_at": "1644731240.5280757"}, "16": {"user_id": "4", "device_id": "3", "assigned_by": "3", "assigned_at": "1644731258.547831"}, "17": {"user_id": "4", "device_id": "5", "assigned_by": "3", "assigned_at": "1644731285.242353"}, "18": {"user_id": "4", "device_id": "3", "assigned_by": "3", "assigned_at": "1644731296.9164596"}, "19": {"user_id": "5", "device_id": "3", "assigned_by": "4", "assigned_at": "1644731387.2701037"}, "20": {"user_id": "4", "device_id": "3", "assigned_by": "3", "assigned_at": "1644731468.283226"}, "21": {"user_id": "6", "device_id": "3", "assigned_by": "5", "assigned_at": "1644731609.0074632"}, "22": {"user_id": "5", "device_id": "3", "assigned_by": "4", "assigned_at": "1644731854.578424"}, "23": {"user_id": "3", "device_id": "1", "assigned_by": "2", "assigned_at": "1644731877.3139257"}, "24": {"user_id": "5", "device_id": "3", "assigned_by": "4", "assigned_at": "1644731899.8514574"}, "25": {"user_id": "1234", "device_id": "3", "assigned_by": "12", "assigned_at": "1644780465.1254942"}, "26": {"user_id": "1234", "device_id": "3", "assigned_by": "12", "assigned_at": "1644780495.3136973"}, "27": {"user_id": "1234", "device_id": "3", "assigned_by": "12", "assigned_at": "1644780516.3238275"}}
