+++ db/device_measurements.json
@@ -1,1 +1,1 @@
-{"0": {"user_id": "", "device_id": "0", "assignment_id": "3", "measurement": "{\"timestamp\": \"1234\", \"temperature\": \"100\"}", "created_at": "1644724935.1757414"}, "1": {"user_id": "", "device_id": "1", "assignment_id": "11", "measurement": "{\"timestamp\": \"1234\", \"blood_pressure\": {\"systolic\": \"120\", \"diastolic\": \"80\"}}", "created_at": "1644727308.5133731"}, "2": {"user_id": "", "device_id": "1", "assignment_id": "12", "measurement": "{\"timestamp\": \"1234\", \"blood_pressure\": {\"systolic\": \"120\", \"diastolic\": \"80\"}}", "created_at": "1644727690.8308377"}, "3": {"user_id": "", "device_id": "1", "assignment_id": "13", "measurement": "{\"timestamp\": \"1234\", \"blood_pressure\": {\"systolic\": \"120\", \"diastolic\": \"80\"}}", "created_at": "1644727950.2580276"}}
+{ "0": {"device_id": "1", "assignment_id": "9", "measurement": {"blood_pressure": "{'systolic': '129', 'diastolic': '78'}", "timestamp": 1644787317.9548879}, "created_at": "1644787317.957279"}}
